chore(Displacement) Add E2E tests
- I also had to handle the bad request error (400) that should occur
when one tries to retrieve a displacement that does not exist. This is
troublesome because in order to do that I had to manually try catch for
the error instead of relying on OpenAPI-backend catching it for
me...I'll try to improve that, or at least make it async with a Promise.